fix: 클라이언트 버튼 컴포넌트 분리
