+++ app/[locale]/academics/components/scholarship/ScholarshipDetail.tsx
@@ -2,7 +2,8 @@
 
 import { deleteScholarshipAction } from '@/actions/academics';
 import { Scholarship, StudentType } from '@/apis/types/academics';
-import { DeleteButton, EditButton } from '@/components/common/Buttons';
+import { EditButton } from '@/components/common/Buttons';
+import { DeleteButton } from '@/components/common/ClientButtons';
 import LoginVisible from '@/components/common/LoginVisible';
 import HTMLViewer from '@/components/form/html/HTMLViewer';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
