+++ app/[locale]/academics/components/timeline/TimelineViewer.tsx
@@ -5,7 +5,8 @@ import { useReducer, useState } from 'react';
 import { TimelineContent } from '@/apis/types/academics';
 import { Attachment } from '@/apis/types/attachment';
 import Attachments from '@/components/common/Attachments';
-import { DeleteButton, EditButton } from '@/components/common/Buttons';
+import { EditButton } from '@/components/common/Buttons';
+import { DeleteButton } from '@/components/common/ClientButtons';
 import LoginVisible from '@/components/common/LoginVisible';
 import HTMLViewer from '@/components/form/html/HTMLViewer';
 import { Link, usePathname } from '@/i18n/routing';
