+++ app/[locale]/community/council/meeting-minute/MinutePageContent.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react';
 
 import { deleteMinuteAction } from '@/actions/council';
 import { CouncilMeetingMinute } from '@/apis/types/council';
-import { DeleteButton, EditButton } from '@/components/common/Buttons';
+import { EditButton } from '@/components/common/Buttons';
 import LoginVisible from '@/components/common/LoginVisible';
 import Timeline from '@/components/common/Timeline';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
@@ -14,6 +14,7 @@ import { getPath } from '@/utils/page';
 import { handleServerResponse } from '@/utils/serverActionError';
 
 import CouncilAttachment from '../components/CouncilAttachments';
+import { DeleteButton } from '@/components/common/ClientButtons';
 
 const minutePath = getPath(councilMinute);
