+++ app/[locale]/people/faculty/[id]/FacultyMemberPageContent.tsx
@@ -4,7 +4,8 @@ import { useTranslations } from 'next-intl';
 
 import { deleteFacultyAction } from '@/actions/people';
 import { Faculty } from '@/apis/types/people';
-import { DeleteButton, EditButton } from '@/components/common/Buttons';
+import { EditButton } from '@/components/common/Buttons';
+import { DeleteButton } from '@/components/common/ClientButtons';
 import LoginVisible from '@/components/common/LoginVisible';
 import { CurvedHorizontalSmallNode } from '@/components/common/Nodes';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
