+++ components/common/Buttons.tsx
@@ -1,13 +1,6 @@
-'use client';
-
 import { MouseEventHandler } from 'react';
-import { useFormStatus } from 'react-dom';
 
 import { Link } from '@/i18n/routing';
-import useModal from '@/utils/hooks/useModal';
-import { CustomError } from '@/utils/serverActionError';
-
-import AlertModal from '../modal/AlertModal';
 
 export function GrayButton({
   title,
@@ -85,21 +78,6 @@ export function CreateButton({ href }: { href: string }) {
   );
 }
 
-export function DeleteButton({ onDelete }: { onDelete: () => Promise<CustomError | void> }) {
-  const { openModal } = useModal();
-
-  return (
-    <GrayButton
-      title="삭제"
-      onClick={() =>
-        openModal(
-          <AlertModal message="삭제하시겠습니까?" confirmText="삭제" onConfirm={onDelete} />,
-        )
-      }
-    />
-  );
-}
-
 export function EditButton({ href }: { href: string }) {
   return (
     <Link href={href} className="inline-block">
@@ -107,29 +85,3 @@ export function EditButton({ href }: { href: string }) {
     </Link>
   );
 }
-
-export const ConfirmButton = ({
-  title,
-  ref,
-  onClick,
-}: {
-  title: string;
-  onClick?: MouseEventHandler<HTMLButtonElement>;
-  ref?: React.RefObject<HTMLButtonElement | null>;
-}) => {
-  const { pending } = useFormStatus();
-
-  return (
-    <button
-      className={`ml-2.5 h-[2.1875rem] rounded-[.0625rem] bg-neutral-700 px-[.875rem] py-[.3125rem] text-md font-medium leading-[1.5rem] text-white hover:bg-neutral-500 ${
-        pending && 'opacity-30'
-      }`}
-      disabled={pending}
-      type="submit"
-      onClick={onClick}
-      ref={ref}
-    >
-      {title}
-    </button>
-  );
-};
